+++ routers/api/v1/repo/repo.go
@@ -34,9 +34,9 @@ import (
 	actions_service "code.gitea.io/gitea/services/actions"
 	"code.gitea.io/gitea/services/context"
 	"code.gitea.io/gitea/services/convert"
+	feed_service "code.gitea.io/gitea/services/feed"
 	"code.gitea.io/gitea/services/issue"
 	repo_service "code.gitea.io/gitea/services/repository"
-	feed_service "code.gitea.io/gitea/services/feed"
 )
 
 // Search repositories via options
